// src/generated/prisma/enums.ts
export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  PRESIDENT: 'PRESIDENT',
  SECRETARY: 'SECRETARY',
  DIRECTOR: 'DIRECTOR',
  BOARD_MEMBER: 'BOARD_MEMBER',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EventStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  CORRECTION_REQUIRED: 'CORRECTION_REQUIRED',
  APPROVED: 'APPROVED',
  ARCHIVED: 'ARCHIVED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]